TRAN-16: Changed files name, updated README and added engines to package.json
