+++ package.json
@@ -156,5 +156,9 @@
     "autoprefixer": "^10.4.8",
     "postcss": "^8.4.16",
     "tailwindcss": "^3.1.8"
+  },
+   "engines": {
+    "node": "18.7.0",
+    "npm": "8.15.0"
   }
 }
