+++ src/components/header/HeaderMenu.tsx
@@ -1,8 +1,8 @@
 import { useMemo } from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import classNames from 'classnames';
-import { Paths } from '../../routes/types';
-import { menuOptions } from './types';
+import { Paths } from '../../routes/paths';
+import { menuOptions } from './menuOptions';
 
 const linkStyle = 'flex h-full items-center px-5 hover:bg-transit-green';
