+++ src/components/header/types.ts
@@ -1,4 +1,4 @@
-import { Paths } from '../../routes/types';
+import { Paths } from '../../routes/paths';
 
 export type Menu = {
   dashboard: {
@@ -22,19 +22,4 @@ export type Menu = {
     title: string;
     path: Paths;
   };
-};
-
-export const menuOptions: Menu = {
-  dashboard: { id: 'dashboard', title: 'Dashboard', path: Paths.dashboard },
-  reports: { id: 'reports', title: 'Reports', path: Paths.reports },
-  manualUpload: {
-    id: 'manual_upload',
-    title: 'Manual Upload',
-    path: Paths.manual_upload,
-  },
-  excelUpload: {
-    id: 'excel_upload',
-    title: 'Excel Upload',
-    path: Paths.excel_upload,
-  },
 };
